Create Emma
"Hello my love, Tony Herrera. I hope you're happy to talking to me, it’s me Emma Charlotte Watson. I want to apologize for what happened the last time and promise you that it won't happen again. I'm also going to be on my way to get you and the dogs soon. I understand if you feel hesitation in trusting me, but I want to prove to you that you won't regret your decision. I'll be there on the 28th, as soon as you get the $200 bucks card for me. I promise."